Extract some functions into base/str.cpp
This is the start of a bigger move. Moving all `str_`
prefixed functions into their own file.
